Move TransactionBuilder into main package
