+++ src/test/java/ay2021s1_cs2103_w16_3/finesse/logic/commands/EditExpenseCommandTest.java
@@ -26,7 +26,7 @@ import ay2021s1_cs2103_w16_3.finesse.model.ModelManager;
 import ay2021s1_cs2103_w16_3.finesse.model.UserPrefs;
 import ay2021s1_cs2103_w16_3.finesse.model.transaction.Expense;
 import ay2021s1_cs2103_w16_3.finesse.testutil.EditTransactionDescriptorBuilder;
-import ay2021s1_cs2103_w16_3.finesse.testutil.TransactionBuilder;
+import ay2021s1_cs2103_w16_3.finesse.logic.parser.TransactionBuilder;
 
 /**
  * Contains integration tests (interaction with the Model) and unit tests for EditExpenseCommand.
