+++ src/test/java/ay2021s1_cs2103_w16_3/finesse/model/transaction/predicates/TitleContainsKeyphrasesPredicateTest.java
@@ -9,7 +9,7 @@ import java.util.List;
 
 import org.junit.jupiter.api.Test;
 
-import ay2021s1_cs2103_w16_3.finesse.testutil.TransactionBuilder;
+import ay2021s1_cs2103_w16_3.finesse.logic.parser.TransactionBuilder;
 
 public class TitleContainsKeyphrasesPredicateTest {
