+++ src/test/java/ay2021s1_cs2103_w16_3/finesse/storage/JsonAdaptedIncomeTest.java
@@ -16,7 +16,7 @@ import ay2021s1_cs2103_w16_3.finesse.model.transaction.Amount;
 import ay2021s1_cs2103_w16_3.finesse.model.transaction.Date;
 import ay2021s1_cs2103_w16_3.finesse.model.transaction.Income;
 import ay2021s1_cs2103_w16_3.finesse.model.transaction.Title;
-import ay2021s1_cs2103_w16_3.finesse.testutil.TransactionBuilder;
+import ay2021s1_cs2103_w16_3.finesse.logic.parser.TransactionBuilder;
 
 public class JsonAdaptedIncomeTest {
     private static final String INVALID_TITLE = "R\u2416chel";
