+++ src/test/java/ay2021s1_cs2103_w16_3/finesse/storage/JsonFinanceTrackerStorageTest.java
@@ -18,7 +18,7 @@ import org.junit.jupiter.api.io.TempDir;
 import ay2021s1_cs2103_w16_3.finesse.commons.exceptions.DataConversionException;
 import ay2021s1_cs2103_w16_3.finesse.model.FinanceTracker;
 import ay2021s1_cs2103_w16_3.finesse.model.ReadOnlyFinanceTracker;
-import ay2021s1_cs2103_w16_3.finesse.testutil.TransactionBuilder;
+import ay2021s1_cs2103_w16_3.finesse.logic.parser.TransactionBuilder;
 
 public class JsonFinanceTrackerStorageTest {
     private static final Path TEST_DATA_FOLDER = Paths.get("src", "test", "data", "JsonFinanceTrackerStorageTest");
